+++ Ship.hpp
@@ -13,8 +13,8 @@ class Ship
 
 public:
     Ship::Ship();
-    Ship::Ship(int id, unsigned int speed, unsigned int maxCrew);
-    Ship::Ship(int id, const std::string& name, unsigned int speed, unsigned int maxCrew, unsigned int capacity);
+    Ship::Ship(int id, size_t speed, size_t maxCrew);
+    Ship::Ship(int id, const std::string& name, size_t speed, size_t maxCrew, size_t capacity);
 
     void setName(const std::string& name);
 };
